Feat: Add basics for Bourrin elroy mode
Pellets scene now emits 3 signals for when a pellet is picked up: Any pellet, normal and power. Thus, other nodes don't have to dive into the children of Pellets to listen to the signals.

Pickables order in the node hierarchy of World has been changed since these 3 signals didn't work before that change.

BourrinElroyMode is a component to attach to EnemyAIBourrin